requirements.txt
requests==2.25.1
python-dotenv==0.19.2
pymongo[srv]
selenium==4.1.0
webdriver-manager==3.5.2